+++ src/main.rs
@@ -1,316 +1,23 @@
 extern crate crypto;
-extern crate swc_common;
-extern crate swc_ecma_parser;
 use clap::{App, Arg};
 use crypto::digest::Digest;
 use crypto::md5::Md5;
 use ignore::overrides::OverrideBuilder;
 use ignore::WalkBuilder;
-use serde::{Deserialize, Serialize};
 use serde_json;
 use std::cmp::min;
 use std::collections::HashMap;
 use std::format;
 use std::fs::{read_to_string, File};
 use std::io::prelude::*;
-use std::path::Path;
 use std::ffi::OsStr;
-use swc_common::{
-  errors::{ColorConfig, Handler},
-  sync::Lrc,
-  BytePos, FileName, SourceMap, Span,
-};
-use swc_ecma_parser::{lexer::Lexer, Capturing, EsConfig, Parser, StringInput, Syntax, TsConfig};
 
-#[derive(Debug, Clone, Deserialize, Serialize)]
-struct CloneLoc {
-  source_id: String,
-  fragement: Option<String>,
-  lo: BytePos,
-  hi: BytePos,
-}
-
-impl CloneLoc {
-  fn new(source_id: String, lo: BytePos, hi: BytePos) -> CloneLoc {
-    CloneLoc {
-      source_id,
-      lo,
-      hi,
-      fragement: None,
-    }
-  }
-}
-
-#[derive(Debug, Clone, Deserialize, Serialize)]
-struct Clone {
-  duplication_a: CloneLoc,
-  duplication_b: CloneLoc,
-}
-
-impl Clone {
-  fn enlarge(&mut self, a_hi: BytePos, b_hi: BytePos) {
-    self.duplication_a.hi = a_hi;
-    self.duplication_b.hi = b_hi;
-  }
-  fn fragement_a(&mut self, fragement_a: String) {
-    self.duplication_a.fragement = Some(fragement_a);
-  }
-  fn fragement_b(&mut self, fragement_b: String) {
-    self.duplication_b.fragement = Some(fragement_b);
-  }
-}
-
-#[derive(Debug, Clone)]
-struct TokenItemValue {
-  id: String,
-  source_id: String,
-  start: Option<Span>,
-  end: Option<Span>,
-}
-
-#[derive(Debug, Clone)]
-struct TokenItem {
-  done: bool,
-  value: TokenItemValue,
-}
-
-struct TokenMap {
-  tokens: std::vec::Vec<swc_ecma_parser::token::TokenAndSpan>,
-  str: String,
-  source_id: String,
-  position: usize,
-  min_token: usize,
-}
-
-impl TokenMap {
-  fn console(&self) {
-    println!("Tokens: {:?}", self.str);
-  }
-  fn substring(&self, start: usize, end: usize) -> &str {
-    &self.str[start..end]
-  }
-  // token length
-  fn size(&self) -> usize {
-    self.tokens.len()
-  }
-  // token hash map length
-  fn len(&self) -> usize {
-    self.str.len()
-  }
-  fn get(&self, index: usize) -> std::option::Option<&swc_ecma_parser::token::TokenAndSpan> {
-    self.tokens.get(index)
-  }
-  fn next(&mut self) -> TokenItem {
-    let istart = min(self.position, self.size() - 1);
-    let iend = min(self.position + self.min_token, self.size() - 1);
-    let start = istart * 32;
-    let end = iend * 32;
-    let start_loc = match self.get(istart) {
-      Some(item) => Some(item.span),
-      _ => None,
-    };
-    let end_loc = match self.get(iend) {
-      Some(item) => Some(item.span),
-      _ => None,
-    };
-    let value = TokenItemValue {
-      id: self.substring(start, end).to_string(),
-      start: start_loc,
-      source_id: self.source_id.clone(),
-      end: end_loc,
-    };
-    let mut last_pos = 1;
-    if self.size() > self.min_token {
-      last_pos = self.size() - self.min_token;
-    }
-    if self.position < last_pos {
-      self.position = self.position + 1;
-      TokenItem { done: false, value }
-    } else {
-      TokenItem { done: true, value }
-    }
-  }
-}
-
-fn detect(
-  tokenmap: &mut TokenMap,
-  store: &mut HashMap<String, TokenItem>,
-  clones: &mut Vec<Clone>,
-) {
-  let mut saved: Option<BytePos> = None;
-  let mut clone: Option<Clone> = None;
-  loop {
-    let item = tokenmap.next();
-    let hi = item.value.end.unwrap().hi;
-    let value = store.get(&item.value.id);
-    let done = item.done;
-    match value {
-      // code frame in store
-      Some(v) => {
-        saved = Some(v.value.end.unwrap().hi);
-        match clone {
-          // clone found first time
-          None => {
-            let duplication_a_lo = match v.value.start {
-              Some(item) => item.lo,
-              _ => BytePos(0),
-            };
-            let duplication_a_hi = match v.value.end {
-              Some(item) => item.hi,
-              _ => BytePos(0),
-            };
-            let duplication_b_lo = match item.value.start {
-              Some(item) => item.lo,
-              _ => BytePos(0),
-            };
-            let duplication_b_hi = match item.value.end {
-              Some(item) => item.hi,
-              _ => BytePos(0),
-            };
-            let duplication_a = CloneLoc::new(
-              v.value.source_id.to_string(),
-              duplication_a_lo,
-              duplication_a_hi,
-            );
-            let duplication_b =
-              CloneLoc::new(item.value.source_id, duplication_b_lo, duplication_b_hi);
-            clone = Some(Clone {
-              duplication_a,
-              duplication_b,
-            });
-          }
-          Some(_) => (),
-        }
-      }
-      // code frame not in store
-      _ => {
-        // save clone
-        match clone {
-          Some(item) => {
-            clones.push(item.clone());
-          }
-          _ => (),
-        }
-        // empty clone
-        clone = None;
-        // set value
-        store.insert(item.value.id.to_string(), item.clone());
-      }
-    }
-    if done == true {
-      // save clone
-      match clone {
-        Some(item) => {
-          clones.push(item.clone());
-        }
-        _ => (),
-      }
-      break;
-    } else {
-      if let Some(ref mut c) = clone {
-        if saved.is_some() {
-          c.enlarge(saved.unwrap(), hi);
-        }
-      }
-    }
-  }
-}
-
-fn tokensize_with_str(input: String) -> std::vec::Vec<swc_ecma_parser::token::TokenAndSpan> {
-  let cm: Lrc<SourceMap> = Default::default();
-  let handler = Handler::with_tty_emitter(ColorConfig::Auto, true, false, Some(cm.clone()));
-  let fm = cm.new_source_file(FileName::Custom("test.js".into()), input);
-
-  let lexer = Lexer::new(
-    Syntax::Es(Default::default()),
-    Default::default(),
-    StringInput::from(&*fm),
-    None,
-  );
-
-  let capturing = Capturing::new(lexer);
-
-  let mut parser = Parser::new_from(capturing);
-
-  for e in parser.take_errors() {
-    e.into_diagnostic(&handler).emit();
-  }
-
-  let _module = parser
-    .parse_module()
-    .map_err(|e| e.into_diagnostic(&handler).emit())
-    .expect("Failed to parse module.");
-
-  let tokens = parser.input().take();
-  tokens
-}
-
-fn tokensize_with_path(filepath: &Path) -> std::vec::Vec<swc_ecma_parser::token::TokenAndSpan> {
-  let cm: Lrc<SourceMap> = Default::default();
-  let handler = Handler::with_tty_emitter(ColorConfig::Auto, true, false, Some(cm.clone()));
-
-  let fm = cm
-    .load_file(filepath)
-    .expect(&format!("failed to load {}", filepath.display()));
-
-  let lexer;
-
-  if filepath.ends_with("jsx") || filepath.ends_with("js") {
-    lexer = Lexer::new(
-      Syntax::Es(EsConfig {
-        jsx: true,
-        class_private_props: true,
-        class_private_methods: true,
-        class_props: true,
-        fn_bind: true,
-        decorators: true,
-        decorators_before_export: true,
-        export_default_from: true,
-        dynamic_import: true,
-        export_namespace_from: true,
-        import_assertions: true,
-        import_meta: true,
-        top_level_await: true,
-        nullish_coalescing: false,
-        num_sep: false,
-        optional_chaining: false,
-      }),
-      Default::default(),
-      StringInput::from(&*fm),
-      None,
-    );
-  } else {
-    lexer = Lexer::new(
-      Syntax::Typescript(TsConfig {
-        tsx: true,
-        dts: false,
-        decorators: true,
-        dynamic_import: true,
-        import_assertions: true,
-        no_early_errors: true,
-      }),
-      Default::default(),
-      StringInput::from(&*fm),
-      None,
-    );
-  }
-
-  let capturing = Capturing::new(lexer);
-
-  let mut parser = Parser::new_from(capturing);
-
-  for e in parser.take_errors() {
-    e.into_diagnostic(&handler).emit();
-  }
-
-  let _module = parser
-    .parse_module()
-    .map_err(|e| e.into_diagnostic(&handler).emit())
-    .expect("Failed to parse module.");
-
-  let tokens = parser.input().take();
-  tokens
-}
+mod tokenmap;
+use tokenmap::{TokenMap, TokenItem, Clone};
+mod parse;
+use parse::{tokensize_with_path};
+mod detect;
+use detect::detect;
 
 fn save(clones: &Vec<Clone>) -> std::io::Result<()> {
   let content = serde_json::to_string(clones)?;
